[⚙️ chore] 인증하기,인증내역 api 파일 분리 #64
