+++ src/lib/axios/verification/post/api.ts
@@ -1,4 +1,4 @@
-import axiosInstance from '../instance';
+import axiosInstance from '../../instance';
 
 export const postMyVerificationApi = async (
   groupId: string,
@@ -31,15 +31,4 @@ export const getQuizApi = async (
     console.error(`getQuiz API 실패`, error);
     throw error;
   }
-};
-
-export const getMyStampApi = (myChallengeId: number) => {
-  return axiosInstance.get(`/myChallenges/${myChallengeId}`);
-};
-
-export const getVerificationsApi = (groupId: string, date: string) => {
-  // date -> 2024-03-19 00:00:00.0
-  return axiosInstance.get(
-    `/challengeGroups/${groupId}/dates?verificationDate=${date}`,
-  );
 };
